User Dashboard Function | Email Functions

+++ includes/reject-booking-inc.php
@@ -17,21 +17,38 @@
         $timeStamp = $_POST['timeStamp'];
         $userID = $_POST['userID'];
         $listingID = $_POST['listingID'];
+        $hospitalName = $_POST['hospitalName'];
 
-
-        $moveToRejected = "INSERT INTO rejectedreservations (user_id, listing_id, firstname, lastname, date, time, phonenumber, email, concern, specifyconcern, hospitalname, reservationtype, booking_timestamp)
-                        VALUES ('$userID', '$listingID', '$firstname','$lastname','$date','$time', '$contactNumber', ' $emailAdd', '$patientConcern', '$specifyConcern', '$patientConcern', '$reservationType', '$timeStamp')";
+        $moveToRejected = "INSERT INTO rejectedreservations (user_id, listing_id, firstname, lastname, date, time, phonenumber, email, concern, specifyconcern, hospitalName, reservationtype, booking_timestamp)
+                        VALUES ('$userID', '$listingID', '$firstname','$lastname','$date','$time', '$contactNumber', ' $emailAdd', '$patientConcern', '$specifyConcern', '$hospitalName', '$reservationType', '$timeStamp')";
 
         $result = mysqli_query($conn, $moveToRejected) or die(mysqli_error($conn));
         
         $lastID = mysqli_insert_id($conn);
 
-        if ($moveToRejected) {
-            $deleteUserBooking = "DELETE FROM userbooking WHERE ID = $bookingID";
-            $deleteResult = mysqli_query($conn, $deleteUserBooking) or die(mysqli_error($conn));
+        $imageDirectory = array();
 
-            if ($deleteUserBooking) {
-                $query = "DELETE FROM referralfiles WHERE referral_id = $bookingID;";
-                $result = mysqli_query($conn, $query) or die(mysqli_error($conn));
+        if ($moveToRejected) {
+            require_once '../PHPMailer/user-booking-rejected.php';
+            sendEmailConfirmation($conn, $userID, $listingID, $firstname, $lastname, $date, $time, $contactNumber, $emailAdd, $patientConcern, $specifyConcern, $hospitalName, $reservationType, $timeStamp);
+
+            $getImageDir = $conn->query("SELECT * FROM referralfiles WHERE booking_id = $bookingID;") or die($conn->error);
+            while ($row = mysqli_fetch_assoc($getImageDir)) {
+                $imageDirectory[] = $row["file_dir"];
+            }
+
+            foreach ($imageDirectory as $toDelete) {
+                unlink($toDelete);
+            }
+
+            $deleteReferral = "DELETE FROM referralfiles WHERE booking_id = $bookingID;";
+            $result = mysqli_query($conn, $deleteReferral) or die(mysqli_error($conn));
+
+            if ($deleteReferral) {
+                $deleteUserBooking = "DELETE FROM userbooking WHERE ID = $bookingID";
+                $deleteResult = mysqli_query($conn, $deleteUserBooking) or die(mysqli_error($conn));
+
             }
         }
+
+
